Add timestamps when archiving DB data
Previously, the variable CSV_NAME in settings.py, which is used by
db_to_csv.py when exporting the database to a CSV file, was just
'tweets.csv'. Since I want to archive a separate version each time I
export, (at least initially), this commit uses the datetime module to
add a timestamp to the CSV filename.

This commit also renames the .db file to include a timestamp as
well. Previously, the variable DB_NAME had been 'tweets.db'. In order
to allow the DB to be renamed when it is archived, this commit changes
this to just 'tweets'.