Update language filtering and add stall warnings
Previously, language had been filtered out through an if check within
the on_status function. This commit takes advantage of the languages
parameter of stream.filter() to do this more efficiently.

In addition, this commit also takes advantage of the stall warnings
parameter of stream.filter() to add notifications if the reader is
falling behind.