Change from SQLite3 to PostgreSQL 10
In preparing to transition my scraper and database to the AWS cloud, I
needed to shift away from SQLite since this is not supported by RDS. I
considered both MySQL and PostgreSQL and chose PostgreSQL because it
seemed to be the nerds' choice (and the MySQL installer messed up both
times I tried it by showing my the initial root user password.)

This commit changes the DB connection string in the settings
file. It also renames what had been twitter_tokens.py to secrets.py
since it now includes the password for the PostgreSQL admin user in
addition to the Twitter API tokens.