Update dependencies
In recreating the environment from this file on my new machine, I had to
manually install the four packages noted in the comment at the bottom.
For three of them, I think listing conda-forge as a channel may solve
the problem. I also realized that I needed to add plotly.