Replace literature files with bibliography
Rather than storing and tracking articles as PDFs in GitHub, I have
moved them to a new "references" subfolder and added a bibliography
instead. This will also act as a commitment device for me to keep an
up-to-date bibliography. Following the convention of the journal Machine
Learning, I have used APA format for citations.