Final report code, figures, and writeups
This commit adds the submitted version of the code, figures, and
capstone project report, with minor additions. The .gitignore file is
updated to ignore PDFs since Markdown files are included. The README now
includes a project overview and replication instructions. The code
folder includes an updated, renamed notebook and the saved spaCy model.
Finally, a new folder containing figures included in the final report is
added.